In all migrations, check for unique keys, indices, foreign keys.

+++ database/migrations/2021_10_25_064208_invites_ref_user_id.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Helpers\HelperMigration;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -14,8 +15,16 @@ class InvitesRefUserId extends Migration
     public function up()
     {
         Schema::table('invites', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id')->nullable()->after('department_id');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+
+            if ( ! Schema::hasColumn('invites', 'user_id') ) {
+                $table->unsignedBigInteger('user_id')->nullable()->after('department_id');
+            }
+
+            $foreignKeys = HelperMigration::listTableForeignKeys('invites');
+            if ( ! in_array('invites_user_id_foreign', $foreignKeys) ) {
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            }
+
         });
     }
 
@@ -27,9 +36,22 @@ class InvitesRefUserId extends Migration
     public function down()
     {
         Schema::table('invites', function (Blueprint $table) {
-            $table->dropForeign('invites_user_id_foreign');
-            $table->dropIndex('invites_user_id_index');
-            $table->dropColumn('user_id');
+
+            $foreignKeys = HelperMigration::listTableForeignKeys('invites');
+            if ( in_array('invites_user_id_foreign', $foreignKeys) ) {
+                $table->dropForeign('invites_user_id_foreign');
+            }
+
+            $sm = Schema::getConnection()->getDoctrineSchemaManager();
+            $doctrineTable = $sm->listTableDetails('invites');
+            if ( $doctrineTable->hasIndex('invites_user_id_index') ) {
+                $table->dropIndex('invites_user_id_index');
+            }
+
+            if ( Schema::hasColumn('invites', 'user_id') ) {
+                $table->dropColumn('user_id');
+            }
+
         });
     }
 }
