+++ database/migrations/2021_10_26_113430_user_has_reward_drop_old_unique_key.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Helpers\HelperMigration;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\DB;
@@ -16,48 +17,12 @@ class UserHasRewardDropOldUniqueKey extends Migration
     {
         Schema::table('user_has_reward', function (Blueprint $table) {
 
-            $table->dropForeign('user_has_reward_user_id_foreign');
-
-//            $table->dropForeign('user_has_reward_user_id_foreign');
-//            $table->dropIndex('user_has_reward_user_id_foreign');
+            $foreignKeys = HelperMigration::listTableForeignKeys('user_has_reward');
+            if ( in_array('user_has_reward_user_id_foreign', $foreignKeys) ) {
+                $table->dropForeign('user_has_reward_user_id_foreign');
+            }
 
         });
-
-//        DB::statement('SET FOREIGN_KEY_CHECKS=0;');
-//        Schema::table('user_has_reward', function (Blueprint $table) {
-//            $table->dropForeign('user_has_reward_user_id_foreign');
-//            $table->dropIndex('user_has_reward_user_id_foreign');
-
-//            $table->dropForeign('user_has_reward_reward_id_foreign');
-//            $table->dropIndex('user_has_reward_reward_id_foreign');
-
-//            $table->dropUnique('user_has_reward_unique');
-//        });
-//        DB::statement('SET FOREIGN_KEY_CHECKS=1;');
-
-//        DB::unprepared("SET SESSION sql_mode = 'ONLY_FULL_GROUP_BY,STRICT_TRANS_TABLES';");
-
-//        DB::unprepared("SET FOREIGN_KEY_CHECKS=0;");
-//        DB::unprepared("ALTER TABLE user_has_reward DROP INDEX user_has_reward_unique;");
-//        DB::unprepared("DROP INDEX user_has_reward_unique ON user_has_reward;");
-//        DB::unprepared("SET FOREIGN_KEY_CHECKS=1;");
-
-//        Schema::table('user_has_reward', function (Blueprint $table) {
-
-//            Schema::disableForeignKeyConstraints();
-
-
-
-//            $table->dropUnique('user_has_reward_unique');
-
-
-//            Schema::enableForeignKeyConstraints();
-
-//            $table->dropUnique(['reward_id', 'user_id', ]);
-//            $table->dropUnique('user_has_reward_unique');
-//            $table->dropUnique('user_has_reward_user_id_reward_id_unique');
-//        });
-
 
     }
 
@@ -70,29 +35,15 @@ class UserHasRewardDropOldUniqueKey extends Migration
     {
         Schema::table('user_has_reward', function (Blueprint $table) {
 
-            $table->foreign('user_id')
-                ->references('id')
-                ->on('users')
-                ->onUpdate('no action')
-                ->onDelete('no action');
+            $foreignKeys = HelperMigration::listTableForeignKeys('user_has_reward');
+            if ( ! in_array('user_has_reward_user_id_foreign', $foreignKeys) ) {
+                $table->foreign('user_id')
+                    ->references('id')
+                    ->on('users')
+                    ->onUpdate('no action')
+                    ->onDelete('no action');
+            }
 
-//            $table->foreign('reward_id')
-//                ->references('id')
-//                ->on('rewards')
-//                ->onUpdate('no action')
-//                ->onDelete('no action');
         });
-
-//        DB::unprepared('ALTER TABLE user_has_reward
-//                              ADD CONSTRAINT user_has_reward_unique
-//                              UNIQUE (user_id,reward_id);');
-
-//        DB::unprepared("SET SESSION sql_mode = 'ONLY_FULL_GROUP_BY,STRICT_TRANS_TABLES';");
-//
-//        Schema::table('user_has_reward', function (Blueprint $table) {
-//            Schema::disableForeignKeyConstraints();
-//            $table->unique(['user_id', 'reward_id'], 'user_has_reward_unique');
-//            Schema::enableForeignKeyConstraints();
-//        });
     }
 }
