+++ database/migrations/2021_10_26_143257_user_has_reward_foreign_key_user_id_two.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Helpers\HelperMigration;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -14,11 +15,16 @@ class UserHasRewardForeignKeyUserIdTwo extends Migration
     public function up()
     {
         Schema::table('user_has_reward', function (Blueprint $table) {
-            $table->foreign('user_id')
-                ->references('id')
-                ->on('users')
-                ->onUpdate('no action')
-                ->onDelete('no action');
+
+            $foreignKeys = HelperMigration::listTableForeignKeys('user_has_reward');
+            if ( ! in_array('user_has_reward_user_id_foreign', $foreignKeys) ) {
+                $table->foreign('user_id')
+                    ->references('id')
+                    ->on('users')
+                    ->onUpdate('no action')
+                    ->onDelete('no action');
+            }
+
         });
     }
 
@@ -30,7 +36,12 @@ class UserHasRewardForeignKeyUserIdTwo extends Migration
     public function down()
     {
         Schema::table('user_has_reward', function (Blueprint $table) {
-            $table->dropForeign('user_has_reward_user_id_foreign');
+
+            $foreignKeys = HelperMigration::listTableForeignKeys('user_has_reward');
+            if ( in_array('user_has_reward_user_id_foreign', $foreignKeys) ) {
+                $table->dropForeign('user_has_reward_user_id_foreign');
+            }
+
         });
     }
 }
